fasterRCNN/requirements.txt
torch
torchvision
numpy
pillow
tqdm
torchmetrics
ultralytics
opencv-python
pandas
matplotlib
seaborn
